refactor(release): reuse demo passport tooling
